PMM-8874 Updated modules for PMM 2.23 (#2067)
* PMM-8874 Updated modules for PMM 2.23

Updated mongodb_exporter, postgres_exporter and toolkit.
Not included yet because we didn't talk about creating a new release for
these exporters.

- proxysql_exporter: release-1.1.2
- rds_exporter: release-0.7.2

I made the branches with the changes. Not need to test because they are
based on the corresponding master braches but there were not official
releases for proxysql_exporter rds_exporter even when we have changes.

* PMM-8874 Using postgres_exporter from master

I've merged PMM-8787 to be able to test directly using master branch so
we can test all tickets together. If there are any problem, we should
revert PMM-8787 merge for PMM 2.23