Make builds work for anyone through disabling private cache by default.
The prior behavior of build-server-rpm used an AWS S3 cache only
accessible with a private key, with no way to disable it. This made it
impossible for anyone without that key to successfully run a build.

The new behavior disables this cache by default, so builds work for
everyone with no modification. It adds a 'cache' variable in the 'vars'
file, controlled by a CACHE environment variable, that allows setting
CACHE to anything besides the value 'false' to enable the private cache.